dcp-815 separate class metadata/datafile.py
